refactor(store): move wake protocols to receivers
Keep durable writes protocol-neutral while sidebar and run consumers own their best-effort datagrams.
