+++ crates/rimz/src/cli/agents_cmd/auto_redeem.rs
@@ -24,7 +24,7 @@ pub(super) fn run_auto_redeem(request: AutoRedeemRequest) -> Result<()> {
         Ok(Some(report)) => {
             append_report(request.kind.as_str(), request.request_id, &report, None);
             if report.reset {
-                let _ = rimz::store::wakeup::wake_sidebars(&runtime);
+                let _ = rimz::sidebar::wakeup::wake_store_delta(&runtime, None, None);
             }
         }
         Ok(None) => {}
