+++ crates/rimz/src/cli/agents_cmd/budget_park.rs
@@ -39,7 +39,7 @@ pub fn run_budget_park(request: BudgetParkRequest) -> Result<()> {
         let (record, wrote) =
             rimz::harness::run::budget_exceeded(store.paths(), &record.run_id, at_cost)?;
         if wrote {
-            let _ = rimz::store::wakeup::wake_run(store.runtime_paths(), &record);
+            let _ = rimz::harness::run_wake::wake_run(store.runtime_paths(), &record);
         }
     }
     interrupted.context("interrupting over-budget agent")
