+++ crates/rimz/src/cli/hooks/lifecycle/context.rs
@@ -155,7 +155,7 @@ pub(super) fn merge_agent_context_sidecars(input: ContextSidecarInput<'_>) {
         );
     }
     if turn_error_updated {
-        let _ = rimz::store::wakeup::wake_sidebars(store.runtime_paths());
+        let _ = rimz::sidebar::wakeup::wake_store_delta(store.runtime_paths(), None, None);
     }
 
     if let Some(context) = decoded.take_observed_context() {
@@ -167,7 +167,7 @@ pub(super) fn merge_agent_context_sidecars(input: ContextSidecarInput<'_>) {
             context.context,
         ) {
             Ok(true) => {
-                let _ = rimz::store::wakeup::wake_sidebars(store.runtime_paths());
+                let _ = rimz::sidebar::wakeup::wake_store_delta(store.runtime_paths(), None, None);
             }
             Ok(false) => {}
             Err(err) => {
@@ -253,7 +253,7 @@ pub(super) fn merge_agent_context_sidecars(input: ContextSidecarInput<'_>) {
     );
     let Some(refresh) = refresh else {
         if locally_priced_cost_updated {
-            let _ = rimz::store::wakeup::wake_sidebars(store.runtime_paths());
+            let _ = rimz::sidebar::wakeup::wake_store_delta(store.runtime_paths(), None, None);
         }
         return;
     };
@@ -271,7 +271,7 @@ pub(super) fn merge_agent_context_sidecars(input: ContextSidecarInput<'_>) {
             "lifecycle: failed to merge local context sidecar",
         );
     } else {
-        let _ = rimz::store::wakeup::wake_sidebars(store.runtime_paths());
+        let _ = rimz::sidebar::wakeup::wake_store_delta(store.runtime_paths(), None, None);
     }
 }
