+++ crates/rimz/src/cli/hooks/lifecycle/reactors.rs
@@ -57,7 +57,7 @@ fn run_wake(ctx: &ReactorCtx<'_>, event: &rimz::agents::LifecycleEvent) {
     let Some(record) = ctx.run_completion else {
         return;
     };
-    if let Err(err) = rimz::store::wakeup::wake_run(ctx.store.runtime_paths(), record) {
+    if let Err(err) = rimz::harness::run_wake::wake_run(ctx.store.runtime_paths(), record) {
         warn!(
             kind = %event.kind,
             agent_id = %event.agent_id,
