+++ crates/rimz/src/harness/run_wake.rs
@@ -38,6 +38,9 @@ pub enum RunWakeErr {
 
     #[error("recv on run wake socket: {0}")]
     Recv(#[source] std::io::Error),
+
+    #[error("serializing run wake frame: {0}")]
+    Serialize(#[from] serde_json::Error),
 }
 
 pub type Result<T> = std::result::Result<T, RunWakeErr>;
@@ -66,8 +69,37 @@ pub enum WakeupFrame {
     },
 }
 
-pub fn run_socket_path(rt: &RuntimePaths, run_id: &RunId) -> PathBuf {
+fn run_socket_path(rt: &RuntimePaths, run_id: &RunId) -> PathBuf {
     rt.sock_dir.join(format!("run.{}.sock", run_id.short()))
+}
+
+/// Send a terminal datagram to the supervised-run waiter. Durable run state
+/// remains authoritative; sender creation and per-target failures are absorbed.
+pub fn wake_run(rt: &RuntimePaths, record: &RunRecord) -> Result<()> {
+    let target = run_socket_path(rt, &record.run_id);
+    if !target.exists() {
+        return Ok(());
+    }
+    let payload = serde_json::to_vec(&WakeupFrame::RunCompleted {
+        workspace_id: record.workspace_id.clone(),
+        run_id: record.run_id.clone(),
+        status: record.status,
+    })?;
+    let sender = match StdUnixDatagram::unbound() {
+        Ok(sender) => sender,
+        Err(error) => {
+            debug!(%error, "run wake: creating sender socket failed");
+            return Ok(());
+        }
+    };
+    if let Err(error) = sender.set_nonblocking(true) {
+        debug!(%error, "run wake: making sender socket non-blocking failed");
+        return Ok(());
+    }
+    if let Err(error) = sender.send_to(&payload, &target) {
+        debug!(?target, %error, "run wake: send_to failed (waiter may have exited)");
+    }
+    Ok(())
 }
 
 /// Bind a per-run datagram socket. Caller owns both the returned
